+++ packages/auth/src/Auth.ts
@@ -4,7 +4,14 @@ import { makeVerify, VerifyEvmOptions, VerifyOptions, VerifySolOptions } from '.
 import {
   VerifyChallengeSolanaResponseAdapter,
   VerifyChallengeEvmResponseAdapter,
+  RequestChallengeEvmResponseAdapter,
+  requestChallengeEvmOperation,
+  RequestChallengeEvmRequest,
+  RequestChallengeSolanaResponseAdapter,
+  requestChallengeSolanaOperation,
+  RequestChallengeSolanaRequest,
 } from '@moralisweb3/common-auth-utils';
+import { OperationResolver } from '@moralisweb3/api-utils';
 
 export const BASE_URL = 'https://authapi.moralis.io';
 
@@ -21,6 +28,23 @@ export class Auth implements ApiModule {
 
   private constructor(private readonly core: Core) {}
 
+  // Client-side compatible operation, structured in a predictable way as defined in the operation
+  // TODO: generate in seperate package "client-evm-auth" (similar to client-evm-auth)
+  public readonly evm = {
+    requestChallengeEvm: (request: RequestChallengeEvmRequest): Promise<RequestChallengeEvmResponseAdapter> => {
+      return new OperationResolver(requestChallengeEvmOperation, this.baseUrl, this.core).fetch(request);
+    },
+  };
+
+  // Client-side compatible operation, structured in a predictable way as defined in the operation
+  // TODO: generate in seperate package "client-evm-auth" (similar to client-evm-auth)
+  public readonly solana = {
+    requestChallengeSol: (request: RequestChallengeSolanaRequest): Promise<RequestChallengeSolanaResponseAdapter> => {
+      return new OperationResolver(requestChallengeSolanaOperation, this.baseUrl, this.core).fetch(request);
+    },
+  };
+
+  // Resolves to requestChallengeEvm/requestChallengeSol depending on provided options (defaults to evm)
   public requestMessage = (options: RequestMessageOptions) => makeRequestMessage(this.core)(options);
 
   // Function overloading to make typescript happy