+++ packages/streams/src/utils/commonNetworkOptions.ts
@@ -1,0 +1,5 @@
+import { StreamNetworkUnion } from '../utils/StreamNetwork';
+
+export interface CommonStreamNetworkOptions {
+  networkType?: StreamNetworkUnion;
+}