+++ packages/client/evmApi/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @moralisweb3/client-evm-api
 
+## 2.11.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/client-api-utils@2.11.1
+  - @moralisweb3/common-core@2.11.1
+  - @moralisweb3/common-evm-utils@2.11.1
+
 ## 2.11.0
 
 ### Patch Changes