+++ packages/common/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @moralisweb3/core
 
+## 2.11.1
+
 ## 2.11.0
 
 ### Patch Changes