+++ packages/evmUtils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @moralisweb3/evm-utils
 
+## 2.11.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/common-evm-utils@2.11.1
+
 ## 2.11.0
 
 ### Patch Changes